# storage: optimize range deletion