--- a/pkg/storage/engine/rocksdb/db.h
+++ b/pkg/storage/engine/rocksdb/db.h
@@ -111,6 +111,9 @@ DBStatus DBGet(DBEngine* db, DBKey key, DBString* value);
 // Deletes the database entry for "key".
 DBStatus DBDelete(DBEngine* db, DBKey key);
 
+// Deletes a range of keys from start (inclusive) to end (exclusive).
+DBStatus DBDeleteRange(DBEngine* db, DBIterator *iter, DBKey start, DBKey end);
+
 // Applies a batch of operations (puts, merges and deletes) to the
 // database atomically and closes the batch. It is only valid to call
 // this function on an engine created by DBNewBatch. If an error is