Optimize deletion of a range of keys by changing the implementation from
iterating over the keys in Go to doing the work in C++. This avoids a
significant number of allocations and cgo calls, though additional
experimentation shows that it is the reduction in cgo calls (not the
reduction in allocations) the provided the bulk of the speedup.

```
name                  old time/op    new time/op    delta
ClearRange_RocksDB-8     424ms ± 4%     226ms ± 3%   -46.72%  (p=0.000 n=10+10)

name                  old speed      new speed      delta
ClearRange_RocksDB-8   158MB/s ± 4%   297MB/s ± 3%   +87.65%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
ClearRange_RocksDB-8      479k ± 0%        0k ± 0%  -100.00%   (p=0.000 n=6+10)
```

This also paves the way for using the RocksDB DeleteRange operation when
it becomes available.

Fixes #12712

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/12745)
<!-- Reviewable:end -->