@tamird 2017-01-06
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> though did not review C++

---

Reviewed 8 of 9 files at r1.
Review status: 8 of 9 files reviewed at latest revision, 1 unresolved discussion, some commit checks failed.

---

*[pkg/storage/engine/bench_test.go, line 601 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-K_pVzgemBXlMuKzqJGs:-K_pVzgf5RiEN8WpOZJ3:b9dfvez) ([raw file](https://github.com/cockroachdb/cockroach/blob/467cda1d43440e48cf5ca039c8278594ebcf05e0/pkg/storage/engine/bench_test.go#L601)):*
> ```Go
> 	const rangeBytes = 64 << 20
> 	const valueBytes = 92
> 	const overhead = 48 // Per key/value overhead (empirically determined)
> ```

can you hoist this constant out into package scope? this is the 4th copy. possibly a bunch more boilerplate duplicated here.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-:-K_pXH4aIXjM9B1xHQ3o:b-k0ndmp)*
<!-- Sent from Reviewable.io -->