@petermattis 2017-01-06
Review status: 8 of 9 files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[pkg/storage/engine/bench_test.go, line 601 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-K_pVzgemBXlMuKzqJGs:-K_pacdYGBhE1-L25_Vy:belf26l) ([raw file](https://github.com/cockroachdb/cockroach/blob/7b4f64ecea6d6cb01d7af845832d5bad973081df/pkg/storage/engine/bench_test.go#L601)):*
<details><summary><i>Previously, tamird (Tamir Duberstein) wrote…</i></summary><blockquote>

can you hoist this constant out into package scope? this is the 4th copy. possibly a bunch more boilerplate duplicated here.
</blockquote></details>

Done. There is a bunch of boilerplate, but I don't think the repetition is harmful. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745)*
<!-- Sent from Reviewable.io -->