@tamird 2017-01-06
Reviewed 2 of 2 files at r2.
Review status: 8 of 9 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745)*
<!-- Sent from Reviewable.io -->