@petermattis 2017-01-07
I added another commit which uses a write-only batch for the replica data deletion in `Replica.destroyDataRaftMuLocked`.