@bdarnell 2017-01-08
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 6 of 9 files at r1, 2 of 2 files at r2, 1 of 1 files at r3.
Review status: all files reviewed at latest revision, 2 unresolved discussions, all commit checks successful.

---

*[pkg/storage/replica.go, line 714 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-K_yhMuu_sTGpS_0b89y:-K_yhMuu_sTGpS_0b89z:b-mc05fe) ([raw file](https://github.com/cockroachdb/cockroach/blob/cf6ad6c86cf844040eea0879b5c4243e51c3f75e/pkg/storage/replica.go#L714)):*
> ```Go
> 	defer batch.Close()
> 
> 	for _, keyRange := range makeAllKeyRanges(r.Desc()) {
> ```

Keep the old comment about using `r.Desc()` instead of `consistentDesc`.

---

*[pkg/storage/engine/rocksdb/db.cc, line 1919 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-K_yjKutbFVIwPm8wLao:-K_yjKutbFVIwPm8wLap:bomo6ex) ([raw file](https://github.com/cockroachdb/cockroach/blob/cf6ad6c86cf844040eea0879b5c4243e51c3f75e/pkg/storage/engine/rocksdb/db.cc#L1919)):*
> ```cpp
>   opts.prefix_same_as_start = prefix;
>   opts.total_order_seek = !prefix;
>   iter->rep.reset(rep->NewIterator(opts));
> ```

This implementation of `NewIter` is safe only if it is the first thing done on the `WriteOnlyBatch`. If something else had been written earlier, the iterator wouldn't see it. Maybe it would be less error-prone if `Engine.ClearRange` took an iterator as an argument to make it explicit when the iterator is coming from a different db view than the batch. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-:-K_yk-e9B9j0S5e1KDL9:bnfp4nl)*
<!-- Sent from Reviewable.io -->