@petermattis 2017-01-09
Review status: 0 of 9 files reviewed at latest revision, 2 unresolved discussions.

---

*[pkg/storage/replica.go, line 714 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-K_yhMuu_sTGpS_0b89y:-Ka-rurMnPJvv36JqDIa:bmmihr9) ([raw file](https://github.com/cockroachdb/cockroach/blob/cf6ad6c86cf844040eea0879b5c4243e51c3f75e/pkg/storage/replica.go#L714)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

Keep the old comment about using `r.Desc()` instead of `consistentDesc`.
</blockquote></details>

Oops, didn't mean to remove that. Re-added.

---

*[pkg/storage/engine/rocksdb/db.cc, line 1919 at r3](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745#-K_yjKutbFVIwPm8wLao:-Ka-s1rjssW6c6lKZ_-X:b-aqk2qw) ([raw file](https://github.com/cockroachdb/cockroach/blob/cf6ad6c86cf844040eea0879b5c4243e51c3f75e/pkg/storage/engine/rocksdb/db.cc#L1919)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

This implementation of `NewIter` is safe only if it is the first thing done on the `WriteOnlyBatch`. If something else had been written earlier, the iterator wouldn't see it. Maybe it would be less error-prone if `Engine.ClearRange` took an iterator as an argument to make it explicit when the iterator is coming from a different db view than the batch. 
</blockquote></details>

Heh, that's the way I originally had it, but removed the iterator to make it compatible with the RocksDB `DeleteRange` operation. I'll add the iterator back for now. It's easy enough to remove when `DeleteRange` arrives.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/12745)*
<!-- Sent from Reviewable.io -->